Clean up codebase
Suppresses some warnings in certain places. test_to_digits was corrected for unsigned overflow by checking if i is less than the size of the array that it iterates over since it will overflow and become greater than the array size.

Fixed some undefined behavior where clear_bit was returning nothing and sigbits_tree was setting the value of key equal to the result of clear_bit. Modified clear_bit to return the value of set_bit as it should have been.