Refactor bitmask and bitmaskv into mask and filter functions.
Much more useful and has better names. Mask creates a mersenne number for masking a set of bits whereas filter can be used for more advanced bit manipulation and masks a set of bits in the value which can itself be used to mask bits corresponding to the original value but only those bits which survive the logical AND of value and the mask. Both preserve bit positions. Both support using signed integer shift operands.