Fix simple allocator API
The simple C allocator was lacking a means to modify the nature of read/write operations to a memory object. This addition perfects the API further to permit changing the memory object from STATE to BEHAVIOR or vice versa; and between any of READ (only), WRITE (only), READ_WRITE, and NOREAD_NOWRITE permissions via m_transmute in memory.h.