Design implementation of m-ary tree (draft)
A more or less complete design of the m-ary tree implementation. I decided this would be more useful to implement and then from there I can make specialized tree structures if I wish. Unfortunately, as C's type system does not work with bits as the atomic unit, it is difficult to implement a pretty interface without the polymorphic constructs necessary to make it work out nicely at runtime, so that is deferred to the AD port in the future.