Rename m_object to m_context
We're now using the term memory context to refer to memory allocations within the BC memory model. Changed the name to reflect that.

Additionally, added some documentation in this regard. I'm thinking MarkDown is a nice formatting system, so I think I'll be using that for parsing and formatting documentation for the time being, but I'll be extending the capacity of MarkDown to something more suitable for rapidly typing documentation with the possibility of writing general documents with as much freedom as one has in HTML, CSS, and ECMAScript combined.