(Allocating) Refactor and retest ImperfectStackUnitAllocator for Win32
Changed the implementation to provide the starting address of a set of allocations created with allocate_all(). It is then up to the user to track the size in order to ensure correct operating from then on, but this is more of a throw-away implementation as C is only being used until we can use our own language, or when necessary.